sumo código para solucionar el punto 2 optimizando tiempo de ejecución y refactor en q2_memory.py por reutilización de funciones
